variants: remove intermediate variants.h
* Move common part of variants/variants.h to core arduino/Arduino.h
  and remove variants/variants.h.
* Rename ..._pinmap.h to variants.h.
* Remove no needed includes.
