--- variants/CMakeLists.txt
@@ -1,3 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-zephyr_include_directories(.)
-zephyr_include_directories(${BOARD})
