--- variants/arduino_mkrzero/arduino_mkrzero_pinmap.h
@@ -1,11 +1,0 @@
-/*
- * Copyright (c) 2022 Dhruva Gole
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#pragma once
-#include <zephyr/drivers/gpio.h>
-#include <zephyr/kernel.h>
-#include <zephyr/drivers/i2c.h>
-
