+++ variants/arduino_nano_33_ble_sense/variants.h
@@ -1,0 +1,5 @@
+/*
+ * Copyright (c) 2022 Dhruva Gole
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
