+++ variants/cc3220sf_launchxl/variants.h
@@ -6,8 +6,9 @@
 
 /* All the pins that are 100 + x are gpio1 pins and < 100 are in gpio0 */
 #pragma once
-#include <zephyr/drivers/gpio.h>
-#include <zephyr/kernel.h>
-#include <zephyr/drivers/i2c.h>
-#include <zephyr/device.h>
+
+#define LED_BUILTIN 4
+#define YELLOW_LED 3
+#define GREEN_LED 4
+#define RED_LED 19
