--- variants/nrf52840dk_nrf52840/nrf52840dk_nrf52840_pinmap.h
@@ -1,14 +1,0 @@
-/*
- * Copyright (c) 2022 Mike Szczys
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#ifndef __NRF52840DK_NRF52840_PINMAP_H__
-#define __NRF52840DK_NRF52840_PINMAP_H__
-
-#include <zephyr/drivers/gpio.h>
-#include <zephyr/drivers/i2c.h>
-#include <zephyr/kernel.h>
-
-#endif
