+++ variants/nrf52840dk_nrf52840/variants.h
@@ -1,0 +1,5 @@
+/*
+ * Copyright (c) 2022 Mike Szczys
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
